Add files via upload
Customer mileage and actual mileage Data.
The geodesic data package was used to visualize the routes and calculate the total number of paths for comparison with the sensor data. The purpose was to find which part of the trip the sensors received a larger difference between the path difference and the actual value.
folium_map_24 and folium_map_25 as two example is to judge the routes displayed for the two days of April 24 and April 25. Estimate_TRIP_distance_with_geodesic analyzes as well as visualizes the routes for all of them.